Add features postman tests

+++ main-service/src/main/java/ru/practicum/main/comment/service/CommentService.java
@@ -1,10 +1,13 @@
 package ru.practicum.main.comment.service;
 
 import lombok.RequiredArgsConstructor;
+
+import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Sort;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 import ru.practicum.main.comment.dto.CommentDto;
 import ru.practicum.main.comment.dto.NewCommentDto;
@@ -28,6 +31,7 @@ import java.util.List;
 
 @Service
 @RequiredArgsConstructor
+@Transactional(readOnly = true)
 public class CommentService {
     private final CommentRepository commentRepository;
     private final EventService eventService;
@@ -49,6 +53,7 @@ public class CommentService {
                 .stream().map(CommentMapper::toCommentDto).toList();
     }
 
+    @Transactional
     public CommentDto createComment(Long userId, NewCommentDto newCommentDto) {
         UserDto user = userService.get(userId);
         EventDto event = eventService.get(newCommentDto.getEventId());
@@ -61,17 +66,22 @@ public class CommentService {
         nUser.setId(user.getId());
 
         Event nEvent = new Event();
-        nUser.setId(event.getId());
+        nEvent.setId(event.getId());
 
         Comment newComment = new Comment();
         newComment.setCreatedOn(LocalDateTime.now());
-        newComment.setEvent(nEvent);
         newComment.setUser(nUser);
+        newComment.setEvent(nEvent);
         newComment.setText(newCommentDto.getText());
         return CommentMapper.toCommentDto(commentRepository.save(newComment));
     }
 
+    @Transactional
     public void deleteCommentByIdByOwner(Long userId, Long commentId) {
+        if (!commentRepository.existsById(commentId)) {
+            throw new NotFoundException("Комментарий с таким id не найден");
+        }
+
         CommentDto stored = getDtoOrThrow(commentId);
         UserDto user = userService.get(userId);
 
@@ -79,43 +89,48 @@ public class CommentService {
             throw new AccessDeniedException("Удалять комментарий может только автор или администратор");
         }
         commentRepository.deleteById(commentId);
+        commentRepository.flush();
     }
 
+    @Transactional
     public void deleteCommentByIdByAdmin(Long commentId) {
-        getDtoOrThrow(commentId);
+        if (!commentRepository.existsById(commentId)) {
+            throw new NotFoundException("Комментарий с таким id не найден");
+        }
         commentRepository.deleteById(commentId);
+        commentRepository.flush();
     }
 
+    @Transactional
     public CommentDto updateCommentForEvent(Long commentId, Long userId, UpdateCommentDto dto) {
-        CommentDto stored = getDtoOrThrow(commentId);
+        if (!commentRepository.existsById(commentId)) {
+            throw new NotFoundException("Комментарий с таким id не найден");
+        }
+
+        Comment stored = commentRepository.findById(commentId).get();
         UserDto user = userService.get(userId);
 
-        if (!user.getId().equals(stored.getUserId())) {
+        if (!user.getId().equals(stored.getUser().getId())) {
             throw new AccessDeniedException("Обновлять комментарий может только автор или администратор");
         }
 
         if (dto.getText() != null) {
             stored.setText(dto.getText());
-        }
 
-        commentRepository.updateTextById(commentId, dto.getText());
+            User nUser = new User();
+            nUser.setId(user.getId());
 
-        return get(commentId);
-    }
-
-    public CommentDto getCommentByIdForUser(Long userId, Long commentId) {
-        UserDto user = userService.get(userId);
-
-        if (!user.getId().equals(userId)) {
-            throw new AccessDeniedException("Информацию о комментарии может получить только автор");
+            Event nEvent = new Event();
+            nEvent.setId(stored.getEvent().getId());
+            commentRepository.save(stored);
         }
         return get(commentId);
     }
 
-    public List<CommentDto> getAll(LocalDateTime rangeStart, LocalDateTime rangeEnd, int from, int size) {
+    public List<CommentDto> getAll(int from, int size) {
         Pageable pageable = PageRequest.of(from / size, size, Sort.by(Sort.Direction.ASC, "createdOn"));
 
-        List<Comment> comments = commentRepository.findAllByCreatedOnBetween(rangeStart, rangeEnd, pageable);
+        Page<Comment> comments = commentRepository.findAll(pageable);
 
         return comments.stream().map(CommentMapper::toCommentDto).toList();
     }
